requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
mongoengine==0.24.2
pymongo==4.0.2
mongomock==4.1.2
pyee==9.0.4
httpx==0.24.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0